test: split integration tests by domain
