+++ tests/integration/init-test-system.ts
@@ -1,0 +1,257 @@
+import { noop } from "lodash/fp";
+import type { Moment } from "moment";
+import { type CachedMetadata, MetadataCache, type Vault } from "obsidian";
+import type { SListEntry, STask } from "obsidian-dataview";
+import { derived, get, writable } from "svelte/store";
+import { isNotVoid } from "typed-assert";
+import { expect, vi } from "vitest";
+
+import { createUpdateHandler } from "../../src/create-update-handler";
+import { dataviewChange } from "../../src/redux/dataview/dataview-slice";
+import { initialState } from "../../src/redux/global-slice";
+import { createReactor, type RootState } from "../../src/redux/store";
+import type { DataviewFacade } from "../../src/service/dataview-facade";
+import { TransactionWriter } from "../../src/service/diff-writer";
+import { ListPropsParser } from "../../src/service/list-props-parser";
+import type { PeriodicNotes } from "../../src/service/periodic-notes";
+import { VaultFacade } from "../../src/service/vault-facade";
+import type { WorkspaceFacade } from "../../src/service/workspace-facade";
+import {
+  type DayPlannerSettings,
+  defaultSettingsForTests,
+} from "../../src/settings";
+import { isLocal, type LocalTask, type Task } from "../../src/task-types";
+import { useTasks } from "../../src/ui/hooks/use-tasks";
+import { getOneLineSummary } from "../../src/util/task-utils";
+
+import {
+  FakeDataviewFacade,
+  FakeMetadataCache,
+  FakePeriodicNotes,
+  FakeWorkspaceFacade,
+  InMemoryVault,
+  type InMemoryFile,
+} from "../test-utils";
+import { loadMetadataDump } from "./load-metadata-dump";
+
+function initTestServices(props: {
+  inMemoryFiles: InMemoryFile[];
+  inMemoryDailyNotes: { path: string; file: InMemoryFile; date: Moment }[];
+  tasks: STask[];
+  lists: SListEntry[];
+  cachedMetadata: Record<string, CachedMetadata>;
+}) {
+  const { inMemoryFiles, inMemoryDailyNotes, tasks, lists, cachedMetadata } =
+    props;
+
+  const periodicNotes = new FakePeriodicNotes(
+    inMemoryDailyNotes,
+  ) as unknown as PeriodicNotes;
+
+  const dataviewFacade = new FakeDataviewFacade({
+    tasks,
+    lists,
+  }) as unknown as DataviewFacade;
+
+  const metadataCache = new FakeMetadataCache(
+    cachedMetadata,
+  ) as unknown as MetadataCache;
+
+  const vault = new InMemoryVault(inMemoryFiles);
+
+  const getTasksApi = () => {
+    throw new Error("Can't access tasks API inside tests");
+  };
+
+  const vaultFacade = new VaultFacade(vault as unknown as Vault, getTasksApi);
+
+  const transactionWriter = new TransactionWriter(vaultFacade);
+
+  const workspaceFacade =
+    new FakeWorkspaceFacade() as unknown as WorkspaceFacade;
+
+  const listPropsParser = new ListPropsParser(
+    vault as unknown as Vault,
+    metadataCache,
+  );
+
+  return {
+    periodicNotes,
+    dataviewFacade,
+    metadataCache,
+    vault,
+    transactionWriter,
+    workspaceFacade,
+    vaultFacade,
+    listPropsParser,
+  };
+}
+
+export async function setUp(props: {
+  visibleDays: string[];
+  settings?: DayPlannerSettings;
+}) {
+  const { visibleDays, settings = defaultSettingsForTests } = props;
+
+  const { inMemoryFiles, inMemoryDailyNotes, tasks, lists, cachedMetadata } =
+    await loadMetadataDump();
+
+  const {
+    periodicNotes,
+    dataviewFacade,
+    metadataCache,
+    vault,
+    transactionWriter,
+    workspaceFacade,
+    vaultFacade,
+    listPropsParser,
+  } = initTestServices({
+    inMemoryFiles,
+    inMemoryDailyNotes,
+    tasks,
+    lists,
+    cachedMetadata,
+  });
+
+  const visibleDaysStore = writable(visibleDays.map((it) => window.moment(it)));
+  const isOnline = writable(true);
+  const layoutReady = writable(true);
+  const settingsStore = writable(settings);
+  const currentTime = writable(window.moment());
+
+  const onEditCanceled = vi.fn();
+  const onEditConfirmed = vi.fn();
+  const defaultPreloadedStateForTests: Partial<RootState> = {
+    obsidian: {
+      ...initialState,
+      visibleDays,
+    },
+    settings: { settings },
+  };
+
+  const {
+    dispatch,
+    remoteTasks,
+    taskUpdateTrigger,
+    listProps,
+    pointerDateTime,
+  } = createReactor({
+    preloadedState: {
+      ...defaultPreloadedStateForTests,
+      obsidian: {
+        ...initialState,
+        visibleDays,
+      },
+    },
+    dataviewFacade,
+    listPropsParser,
+  });
+
+  inMemoryFiles.forEach(({ path }) => {
+    dispatch(dataviewChange(path));
+  });
+
+  const onUpdate = createUpdateHandler({
+    settings: () => settings,
+    transactionWriter,
+    vaultFacade,
+    periodicNotes,
+    onEditCanceled,
+    onEditConfirmed,
+    getTextInput: () => Promise.resolve("Text input"),
+    getConfirmationInput: () => Promise.resolve(true),
+  });
+
+  const {
+    tasksWithActiveClockProps,
+    getDisplayedTasksWithClocksForTimeline,
+    tasksWithTimeForToday,
+    editContext,
+    newlyStartedTasks,
+  } = useTasks({
+    onUpdate,
+    onEditAborted: () => {},
+    periodicNotes,
+    dataviewFacade,
+    metadataCache,
+    workspaceFacade,
+    isOnline,
+    visibleDays: visibleDaysStore,
+    layoutReady,
+    debouncedTaskUpdateTrigger: taskUpdateTrigger,
+    dataviewChange: taskUpdateTrigger,
+    settingsStore,
+    currentTime,
+    pointerDateTime,
+    remoteTasks,
+    listProps,
+  });
+
+  const allTasks = derived(
+    editContext.dayToDisplayedTasks,
+    ($dayToDisplayedTasks) => {
+      return Object.values($dayToDisplayedTasks).flatMap(
+        ({ withTime, noTime }) => withTime.concat(noTime),
+      );
+    },
+  );
+
+  allTasks.subscribe(noop);
+
+  function moveCursorTo(
+    dateTime: Moment,
+    type: "date" | "dateTime" = "dateTime",
+  ) {
+    pointerDateTime.set({
+      dateTime,
+      type,
+    });
+  }
+
+  function findTask(predicate: (task: Task) => boolean) {
+    const found = get(allTasks).filter(isLocal).find(predicate);
+
+    isNotVoid(found, "Task not found");
+
+    return found;
+  }
+
+  async function findByText(text: string) {
+    let found: LocalTask | undefined;
+
+    await vi.waitFor(() => {
+      found = get(allTasks)
+        .filter(isLocal)
+        .find((it) => getOneLineSummary(it).includes(text));
+
+      expect(found).toBeDefined();
+    });
+
+    return found as LocalTask;
+  }
+
+  await vi.waitFor(() => {
+    const areTasksLoaded = get(allTasks).length > 0;
+    const areClocksLoaded = get(tasksWithActiveClockProps).length > 0;
+
+    expect(areTasksLoaded || areClocksLoaded).toBeTruthy();
+  });
+
+  return {
+    dispatch,
+    tasksWithActiveClockProps,
+    getDisplayedTasksWithClocksForTimeline,
+    tasksWithTimeForToday,
+    editContext,
+    newlyStartedTasks,
+    periodicNotes,
+    moveCursorTo,
+    onUpdate,
+    vault,
+    findTask,
+    findByText,
+    allTasks,
+    transactionWriter,
+    currentTime,
+  };
+}
